% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_helpers.R
\name{tmle3sim_traceback}
\alias{tmle3sim_traceback}
\title{Get traceback for last simulation error}
\usage{
tmle3sim_traceback(call_stack = NULL, cat = TRUE)
}
\arguments{
\item{call_stack}{a call stack argument (defaults to last simulation error)}

\item{cat}{if TRUE, uses cat to print the traceback}
}
\description{
Get traceback for last simulation error
}
